feat: implement DiseaseCaseController and DiseaseCaseService for managing disease cases

+++ backend/src/main/java/mspr/backend/Repository/DiseaseCaseRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import mspr.backend.BO.DiseaseCase;
 
 @Repository
-public interface DiseaseCaseRepository extends JpaRepository<DiseaseCase, Long> {
+public interface DiseaseCaseRepository extends JpaRepository<DiseaseCase, Integer> {
 }
